+++ src/quadtree/morton_key.rs
@@ -31,10 +31,9 @@ impl MortonKey {
     /// Calculate the original point of this hash key.
     /// In practice it is more beneficial to just store the original key if you need to access it
     /// later.
-    #[allow(unused)]
-    pub fn as_point(&self) -> [u16; 2] {
-        let x = Self::reconstruct(self.0) as u16;
-        let y = Self::reconstruct(self.0 >> 1) as u16;
+    pub fn as_point(&self) -> [u32; 2] {
+        let x = Self::reconstruct(self.0);
+        let y = Self::reconstruct(self.0 >> 1);
         [x, y]
     }
